+++ server/udp/binder.go
@@ -25,7 +25,20 @@ func (binder Binder) Start(ws *websocket.Conn, store *service.SessionStore) erro
 	}
 	defer conn.Close()
 
+	// Forcifully close connection (thus end session) if the agent does not
+	// respond to ping.
 	go service.Watch(ws, connTimeout, conn.Close)
+
+	go func() {
+		for {
+			// Agent does not send message to this channel in the current
+			// protocol, but it is required to drain the channel to check
+			// for ping responses.
+			if _, _, err := ws.NextReader(); err != nil {
+				break
+			}
+		}
+	}()
 
 	buf := make([]byte, config.BufferSize)